+++ CHANGELOG.md
@@ -80,6 +80,8 @@
   [#140](https://github.com/cadeyrn/enterprise-policy-generator/issues/140)
 - removed InstallAddonsPermission policy and migrated all data to ExtensionSettings policy in saved configurations, see
   [#168](https://github.com/cadeyrn/enterprise-policy-generator/issues/168)
+- removed FlashPlugin policy as Flash is no longer a thing, see
+  [#218](https://github.com/cadeyrn/enterprise-policy-generator/issues/218)
 
 #### Dependencies